Implement a simple search form
This search form displays fields for choosing options about flights,
then submits it to the same route for search. the controller recognizes
if there are query parameters, and if there are, passes them to a search
method in the model.

Next, the view should try to display those results in a table.